requirements.txt
opencv-python
pillow
numpy
faker
scikit-learn
pandas
streamlit
pytesseract
